+++ python/semantic_kernel/connectors/ai/google/shared.py
@@ -1,0 +1,23 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from semantic_kernel.contents.chat_history import ChatHistory
+from semantic_kernel.contents.utils.author_role import AuthorRole
+from semantic_kernel.exceptions.service_exceptions import ServiceInvalidRequestError
+
+
+def filter_system_message(chat_history: ChatHistory) -> str | None:
+    """Filter the first system message from the chat history.
+
+    If there are multiple system messages, raise an error.
+    If there are no system messages, return None.
+    """
+    if len([message for message in chat_history if message.role == AuthorRole.SYSTEM]) > 1:
+        raise ServiceInvalidRequestError(
+            "Multiple system messages in chat history. Only one system message is expected."
+        )
+
+    for message in chat_history:
+        if message.role == AuthorRole.SYSTEM:
+            return message.content
+
+    return None
